refactor: remove support for mistralai client v0
